reorganize consts and helper to utils.js. Fix ethers.utils method calls
